Consolidation of tests
